+++ src/index.js
@@ -1,22 +1,24 @@
 import MetaMaskOnboarding from '@metamask/onboarding';
-import { createWeb3Modal, defaultConfig } from '@web3modal/ethers5';
 // eslint-disable-next-line camelcase
 import {
   encrypt,
   recoverPersonalSignature,
-  recoverTypedSignatureLegacy,
   recoverTypedSignature,
-  recoverTypedSignature_v4 as recoverTypedSignatureV4,
-} from 'eth-sig-util';
+} from '@metamask/eth-sig-util';
 import { ethers } from 'ethers';
 import { toChecksumAddress } from 'ethereumjs-util';
-import { getPermissionsDisplayString, stringifiableToHex } from './utils';
+import {
+  handleSdkConnect,
+  handleWalletConnect,
+  walletConnect,
+} from './connections';
 import Constants from './constants.json';
 import {
-  NETWORKS_BY_CHAIN_ID,
   ERC20_SAMPLE_CONTRACTS,
   ERC721_SAMPLE_CONTRACTS,
+  NETWORKS_BY_CHAIN_ID,
 } from './onchain-sample-contracts';
+import { getPermissionsDisplayString, stringifiableToHex } from './utils';
 
 const {
   hstBytecode,
@@ -78,7 +80,8 @@ const warningDiv = document.getElementById('warning');
 const onboardButton = document.getElementById('connectButton');
 const getAccountsButton = document.getElementById('getAccounts');
 const getAccountsResult = document.getElementById('getAccountsResult');
-const openConnectModalBtn = document.getElementById('open-connect-modal');
+const walletConnectBtn = document.getElementById('walletConnect');
+const sdkConnectBtn = document.getElementById('sdkConnect');
 
 // Permissions Actions Section
 const requestPermissionsButton = document.getElementById('requestPermissions');
@@ -120,6 +123,7 @@ const revokeButton = document.getElementById('revokeButton');
 const transferTokenInput = document.getElementById('transferTokenInput');
 const transferFromButton = document.getElementById('transferFromButton');
 const nftsStatus = document.getElementById('nftsStatus');
+const erc721TokenAddresses = document.getElementById('erc721TokenAddresses');
 
 // ERC 1155 Section
 
@@ -141,6 +145,7 @@ const revokeERC1155Button = document.getElementById('revokeERC1155Button');
 const watchAssetInput = document.getElementById('watchAssetInput');
 const watchAssetButton = document.getElementById('watchAssetButton');
 const erc1155Status = document.getElementById('erc1155Status');
+const erc1155TokenAddresses = document.getElementById('erc1155TokenAddresses');
 
 // ERC 747 Section
 const eip747ContractAddress = document.getElementById('eip747ContractAddress');
@@ -166,7 +171,7 @@ const transferFromRecipientInput = document.getElementById(
   'transferFromRecipientInput',
 );
 const tokenSymbol = 'TST';
-const tokenAddresses = document.getElementById('tokenAddresses');
+const erc20TokenAddresses = document.getElementById('erc20TokenAddresses');
 const createToken = document.getElementById('createToken');
 const watchAssets = document.getElementById('watchAssets');
 const transferTokens = document.getElementById('transferTokens');
@@ -494,49 +499,6 @@ const initialConnectedButtons = [
   maliciousPermitIntAddress,
 ];
 
-// Buttons that are available after connecting via Wallet Connect
-const walletConnectButtons = [
-  sendButton,
-  personalSign,
-  signTypedData,
-  ethSign,
-  personalSign,
-  signTypedData,
-  signTypedDataV3,
-  signTypedDataV4,
-  signTypedDataV4Batch,
-  signTypedDataV4Queue,
-  signPermit,
-  siwe,
-  siweResources,
-  siweBadDomain,
-  siweBadAccount,
-  siweMalformed,
-  signInvalidType,
-  signEmptyDomain,
-  signExtraDataNotTyped,
-  signInvalidPrimaryType,
-  signNoPrimaryTypeDefined,
-  signInvalidVerifyingContractType,
-  eip747WatchButton,
-  maliciousApprovalButton,
-  maliciousSetApprovalForAll,
-  maliciousERC20TransferButton,
-  maliciousRawEthButton,
-  maliciousPermit,
-  maliciousTradeOrder,
-  maliciousSeaport,
-  sendWithInvalidValue,
-  sendWithInvalidTxType,
-  sendWithInvalidRecipient,
-  mintSepoliaERC20,
-  maliciousSendWithOddHexData,
-  maliciousSendWithoutHexPrefixValue,
-  maliciousApproveERC20WithOddHexData,
-  maliciousPermitHexPaddedChain,
-  maliciousPermitIntAddress,
-];
-
 /**
  * Provider
  */
@@ -548,62 +510,33 @@ let scrollToHandled = false;
 
 const isMetaMaskConnected = () => accounts && accounts.length > 0;
 let isWalletConnectConnected = false;
+let isSdkConnected = false;
 
 // TODO: Need to align with @metamask/onboarding
 const isMetaMaskInstalled = () => provider && provider.isMetaMask;
 
-// test id
-const projectId = 'e6360eaee594162688065f1c70c863b7';
-
-const metadata = {
-  name: 'E2e Test Dapp',
-  description: 'This is the E2e Test Dapp',
-  url: 'https://metamask.github.io/test-dapp/',
-  icons: ['https://avatars.mywebsite.com/'],
+walletConnectBtn.onclick = () => {
+  walletConnect.open();
+  walletConnect.subscribeProvider(() => {
+    handleWalletConnect(
+      'wallet-connect',
+      walletConnectBtn,
+      isWalletConnectConnected,
+    );
+  });
 };
 
-const modal = createWeb3Modal({
-  ethersConfig: defaultConfig({ metadata }),
-  projectId,
-});
+sdkConnectBtn.onclick = async () => {
+  await handleSdkConnect('sdk-connect', sdkConnectBtn, isSdkConnected);
+};
 
-async function handleWalletConnectChange({ isConnected }) {
-  if (isConnected) {
-    provider = modal.getWalletProvider().provider;
-    const providerDetail = {
-      info: {
-        uuid: provider.signer.uri,
-        name: 'wallet-connect',
-        icon: './wallet-connect.svg',
-        rdns: 'io.metamask',
-      },
-      provider,
-    };
-    setActiveProviderDetail(providerDetail);
-    handleNewProviderDetail(providerDetail);
-
-    isWalletConnectConnected = true;
-    updateFormElements();
-    try {
-      const newAccounts = await provider.request({
-        method: 'eth_accounts',
-      });
-      handleNewAccounts(newAccounts);
-    } catch (err) {
-      console.error('Error on init when getting accounts', err);
-    }
-  } else {
-    isWalletConnectConnected = false;
-    openConnectModalBtn.innerText = 'Wallet Connect';
-    handleNewAccounts([]);
-    updateFormElements();
-  }
+export function updateWalletConnectState(isConnected) {
+  isWalletConnectConnected = isConnected;
 }
 
-openConnectModalBtn.onclick = () => {
-  modal.open();
-  modal.subscribeProvider(handleWalletConnectChange);
-};
+export function updateSdkConnectionState(isConnected) {
+  isSdkConnected = isConnected;
+}
 
 const detectEip6963 = () => {
   window.addEventListener('eip6963:announceProvider', (event) => {
@@ -618,7 +551,7 @@ const detectEip6963 = () => {
   window.dispatchEvent(new Event('eip6963:requestProvider'));
 };
 
-const setActiveProviderDetail = async (providerDetail) => {
+export const setActiveProviderDetail = async (providerDetail) => {
   closeProvider();
   provider = providerDetail.provider;
   initializeProvider();
@@ -680,13 +613,28 @@ const existsProviderDetail = (newProviderDetail) => {
   return false;
 };
 
-const handleNewProviderDetail = (newProviderDetail) => {
+export const handleNewProviderDetail = (newProviderDetail) => {
   if (existsProviderDetail(newProviderDetail)) {
     return;
   }
-
   providerDetails.push(newProviderDetail);
+  renderProviderDetails();
+};
 
+export const removeProviderDetail = (name) => {
+  const index = providerDetails.findIndex(
+    (providerDetail) => providerDetail.info.name === name,
+  );
+  if (index === -1) {
+    console.log(`ProviderDetail with name ${name} not found`);
+    return;
+  }
+  providerDetails.splice(index, 1);
+  renderProviderDetails();
+  console.log(`ProviderDetail with name ${name} removed successfully`);
+};
+
+const renderProviderDetails = () => {
   providersDiv.innerHTML = '';
   providerDetails.forEach((providerDetail) => {
     const { info, provider: provider_ } = providerDetail;
@@ -719,7 +667,7 @@ const handleNewProviderDetail = (newProviderDetail) => {
   });
 };
 
-const handleNewAccounts = (newAccounts) => {
+export const handleNewAccounts = (newAccounts) => {
   accounts = newAccounts;
   updateFormElements();
 
@@ -987,7 +935,7 @@ const initializeContracts = () => {
 
 // Must be called after the provider or connect acccounts change
 // Updates form elements content and disabled status
-const updateFormElements = () => {
+export const updateFormElements = () => {
   const accountButtonsDisabled =
     !isMetaMaskInstalled() || !isMetaMaskConnected();
   if (accountButtonsDisabled) {
@@ -996,14 +944,7 @@ const updateFormElements = () => {
     }
     clearDisplayElements();
   }
-  if (
-    isWalletConnectConnected &&
-    activeProviderNameResult.innerText === 'wallet-connect'
-  ) {
-    for (const button of walletConnectButtons) {
-      button.disabled = false;
-    }
-  } else if (isMetaMaskConnected()) {
+  if (isMetaMaskConnected()) {
     for (const button of initialConnectedButtons) {
       button.disabled = false;
     }
@@ -1068,8 +1009,6 @@ const updateOnboardElements = () => {
   }
 
   if (isWalletConnectConnected) {
-    openConnectModalBtn.innerText = 'Wallet Connect - Connected';
-
     if (onboarding) {
       onboarding.stopOnboarding();
     }
@@ -1097,6 +1036,7 @@ const updateContractElements = () => {
     multisigContractStatus.innerHTML = 'Deployed';
     sendMultisigButton.disabled = false;
     // ERC721 Token - NFTs contract
+    erc721TokenAddresses.innerHTML = nftsContract ? nftsContract.address : '';
     nftsStatus.innerHTML = 'Deployed';
     mintButton.disabled = false;
     mintAmountInput.disabled = false;
@@ -1112,6 +1052,9 @@ const updateContractElements = () => {
     watchNFTButtons.innerHTML = '';
 
     // ERC 1155 Multi Token
+    erc1155TokenAddresses.innerHTML = erc1155Contract
+      ? erc1155Contract.address
+      : '';
     erc1155Status.innerHTML = 'Deployed';
     batchMintButton.disabled = false;
     batchMintTokenIds.disabled = false;
@@ -1124,7 +1067,7 @@ const updateContractElements = () => {
     watchAssetInput.disabled = false;
     watchAssetButton.disabled = false;
     // ERC20 Token - Send Tokens
-    tokenAddresses.innerHTML = hstContract ? hstContract.address : '';
+    erc20TokenAddresses.innerHTML = hstContract ? hstContract.address : '';
     watchAssets.disabled = false;
     transferTokens.disabled = false;
     transferFromTokens.disabled = false;
@@ -1314,6 +1257,13 @@ const initializeFormElements = () => {
     console.log(
       `Contract mined! address: ${nftsContract.address} transactionHash: ${nftsContract.deployTransaction.hash}`,
     );
+
+    erc721TokenAddresses.innerHTML = erc721TokenAddresses.innerHTML
+      .concat(', ', nftsContract.address)
+      .split(', ')
+      .filter(Boolean)
+      .join(', ');
+
     nftsStatus.innerHTML = 'Deployed';
     mintButton.disabled = false;
     mintAmountInput.disabled = false;
@@ -1464,6 +1414,12 @@ const initializeFormElements = () => {
     console.log(
       `Contract mined! address: ${erc1155Contract.address} transactionHash: ${erc1155Contract.deployTransaction.hash}`,
     );
+
+    erc1155TokenAddresses.innerHTML = erc1155TokenAddresses.innerHTML
+      .concat(', ', erc1155Contract.address)
+      .split(', ')
+      .filter(Boolean)
+      .join(', ');
 
     erc1155Status.innerHTML = 'Deployed';
     batchTransferTokenIds.disabled = false;
@@ -1807,7 +1763,7 @@ const initializeFormElements = () => {
       );
       await hstContract.deployTransaction.wait();
     } catch (error) {
-      tokenAddresses.innerHTML = 'Creation Failed';
+      erc20TokenAddresses.innerHTML = 'Creation Failed';
       throw error;
     }
 
@@ -1818,7 +1774,7 @@ const initializeFormElements = () => {
     console.log(
       `Contract mined! address: ${hstContract.address} transactionHash: ${hstContract.deployTransaction.hash}`,
     );
-    tokenAddresses.innerHTML = tokenAddresses.innerHTML
+    erc20TokenAddresses.innerHTML = erc20TokenAddresses.innerHTML
       .concat(', ', hstContract.address)
       .split(', ')
       .filter(Boolean)
@@ -1840,7 +1796,7 @@ const initializeFormElements = () => {
   };
 
   watchAssets.onclick = async () => {
-    const contractAddresses = tokenAddresses.innerHTML.split(', ');
+    const contractAddresses = erc20TokenAddresses.innerHTML.split(', ');
 
     const promises = contractAddresses.map((erc20Address) => {
       return provider.request({
@@ -2037,11 +1993,11 @@ const initializeFormElements = () => {
   encryptButton.onclick = () => {
     try {
       ciphertextDisplay.innerText = stringifiableToHex(
-        encrypt(
-          encryptionKeyDisplay.innerText,
-          { data: encryptMessageInput.value },
-          'x25519-xsalsa20-poly1305',
-        ),
+        encrypt({
+          publicKey: encryptionKeyDisplay.innerText,
+          data: encryptMessageInput.value,
+          version: 'x25519-xsalsa20-poly1305',
+        }),
       );
       decryptButton.disabled = false;
     } catch (error) {
@@ -2252,7 +2208,7 @@ const initializeFormElements = () => {
       const sign = personalSignResult.innerHTML;
       const recoveredAddr = recoverPersonalSignature({
         data: msg,
-        sig: sign,
+        signature: sign,
       });
       if (recoveredAddr === from) {
         console.log(`SigUtil Successfully verified signer as ${recoveredAddr}`);
@@ -2331,9 +2287,10 @@ const initializeFormElements = () => {
     try {
       const from = accounts[0];
       const sign = signTypedDataResult.innerHTML;
-      const recoveredAddr = await recoverTypedSignatureLegacy({
+      const recoveredAddr = await recoverTypedSignature({
         data: msgParams,
-        sig: sign,
+        signature: sign,
+        version: 'V1',
       });
       if (toChecksumAddress(recoveredAddr) === toChecksumAddress(from)) {
         console.log(`Successfully verified signer as ${recoveredAddr}`);
@@ -2450,7 +2407,8 @@ const initializeFormElements = () => {
       const sign = signTypedDataV3Result.innerHTML;
       const recoveredAddr = await recoverTypedSignature({
         data: msgParams,
-        sig: sign,
+        signature: sign,
+        version: 'V3',
       });
       if (toChecksumAddress(recoveredAddr) === toChecksumAddress(from)) {
         console.log(`Successfully verified signer as ${recoveredAddr}`);
@@ -2595,9 +2553,10 @@ const initializeFormElements = () => {
     try {
       const from = accounts[0];
       const sign = signTypedDataV4Result.innerHTML;
-      const recoveredAddr = recoverTypedSignatureV4({
+      const recoveredAddr = recoverTypedSignature({
         data: msgParams,
-        sig: sign,
+        signature: sign,
+        version: 'V4',
       });
       if (toChecksumAddress(recoveredAddr) === toChecksumAddress(from)) {
         console.log(`Successfully verified signer as ${recoveredAddr}`);
@@ -2744,9 +2703,10 @@ const initializeFormElements = () => {
     };
     try {
       const sign = signPermitResult.innerHTML;
-      const recoveredAddr = recoverTypedSignatureV4({
+      const recoveredAddr = recoverTypedSignature({
         data: msgParams,
-        sig: sign,
+        signature: sign,
+        version: 'V4',
       });
       if (toChecksumAddress(recoveredAddr) === toChecksumAddress(from)) {
         console.log(`Successfully verified signer as ${recoveredAddr}`);
@@ -3391,6 +3351,7 @@ const setDeeplinks = () => {
 const initialize = async () => {
   setActiveProviderDetailWindowEthereum();
   detectEip6963();
+  setActiveProviderDetail(providerDetails[0]);
   initializeFormElements();
   setDeeplinks();
 };